test(plugin): Phase C.M5 — extract multiclientRpc helper for reuse (0.4.39)
Pure refactor. Pulls the per-client RPC stack (`RpcClientHandle` +
`buildRpcClient`) and the `fs.watch` queue/awaitNext helper
(`watchFor`) out of `tests/integration/multiclient.rpc.test.ts`
into a new `tests/integration/helpers/multiclientRpc.ts` so the
upcoming Phase C tests can reuse the same primitives without
duplicating session setup or notification plumbing.

Why now (atomic):

- M6 (perf.sync.bench.test.ts) needs two RPC clients on a shared
  daemon, indistinguishable from what `multiclient.rpc.test.ts`
  already builds. Cloning the helpers would set up a maintenance
  hazard the first time the watch-queue logic gets touched.
- M9 (sync.e2e.test.ts) likewise.
- Doing the extraction *before* M2/M3 instrumentation lands keeps
  the diffs orthogonal: M2/M3 will only edit production code, M6
  / M9 will only add new test files.

Behavioural contract: `multiclient.rpc.test.ts` is unchanged in
substance. The body of every test (setup, F4 create / modify /
delete cases, teardown) is byte-identical; only the helper
declarations moved out and the imports gained one line.

What ships:

- `tests/integration/helpers/multiclientRpc.ts` (new):
  - `RpcClientHandle { ssh, conn, close() }` interface
  - `buildRpcClient(socketPath, token, label)` — fresh SSH session +
    `establishRpcConnection` over a unix-socket stream
  - `watchFor(client, path)` — subscribes `fs.watch` recursively,
    returns `awaitNext(predicate, timeoutMs?)` + `cleanup()`. Queue
    drains on every notification so a predicate registered after
    the event still resolves.

- `tests/integration/multiclient.rpc.test.ts` (refactor):
  - Imports `buildRpcClient`, `watchFor`, `RpcClientHandle` from
    the new helper instead of declaring them inline.
  - `import type { FsChangedParams }` removed (now used only by the
    helper).
  - Doc comment updated to point at the helper for downstream
    reusers (M6 perf bench, M9 E2E suite).
  - Net change: -104 lines from the test file, +111 lines in the
    helper, +1 import line.

Verification:

- `npx vitest run` — 32 files / 429 tests green (was 429, +0; the
  refactored integration test isn't in the unit config so the count
  is unchanged but the unit suite still passes against the moved
  symbols' transitive imports).
- `npx tsc --noEmit` (custom invocation including `tests/integration`
  since the project tsconfig only includes `src/**/*.ts`) — clean
  on the new helper and the refactored test.
- Integration suite itself isn't run here (Docker isn't available
  on this dev box per memory); the integration CI workflow will
  run the unchanged behavioural assertions against the refactored
  imports.

manifest/package/versions bumped 0.4.38 → 0.4.39.
