+++ plugin/versions.json
@@ -42,5 +42,6 @@
   "0.4.35": "1.4.0",
   "0.4.36": "1.4.0",
   "0.4.37": "1.4.0",
-  "0.4.38": "1.4.0"
+  "0.4.38": "1.4.0",
+  "0.4.39": "1.4.0"
 }
